docs-theme: prep theme for moving docs server
Prepare the docs theme for moving the docs away from zephyrproject.org and on
to a new docs.zephyrproject.org server by adjust header links to not be
relative (e.g., /about changes to zephyrproject.org/about). This change can be
made now since it still works if the docs and main site are on the same
server.)

Also allow "Docs" in breadcrumb to be parameterized in conf.py (default remains
as "Docs") so when we regenerate docs for older releases we can include the
release version in the breadcrumb header.

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>